process: Rename postprocessing to processing
